refator: move logging senative masking configuration to TelemetryCore
